requirements_my.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
basicsr=1.2.0+unknown=dev_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.08.22=haa95532_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py39h59b6b97_0
cryptography=39.0.1=py39h21b164f_2
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
einops=0.6.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2024.3.1=pypi_0
future=0.18.3=pypi_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
google-auth=2.21.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.56.0=pypi_0
h5py=3.9.0=py39hfc34f40_0
hdf5=1.12.1=h51c971a_3
huggingface-hub=0.21.4=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=h59b6b97_46319
joblib=1.2.0=py39haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py39hd77b12b_0
kornia=0.7.0=pypi_0
krb5=1.20.1=h5b6d351_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lmdb=1.4.1=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=py39haa95532_0
matplotlib-base=3.7.2=py39h4ed8f06_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.6=py39hf11a4ad_1
mkl_random=1.2.2=py39hf11a4ad_1
munkres=1.1.4=py_0
natsort=7.1.1=pyhd3eb1b0_0
networkx=3.1=pypi_0
numpy=1.25.0=py39h055cbcc_0
numpy-base=1.25.0=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.74=pypi_0
openssl=3.0.11=h2bbff1b_2
packaging=23.1=py39haa95532_0
pcre=8.45=hd77b12b_0
pillow=9.4.0=py39hd77b12b_0
pip=23.1.2=py39haa95532_0
platformdirs=3.8.1=pypi_0
ply=3.11=py39haa95532_0
protobuf=4.23.4=pypi_0
ptflops=0.7=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.7=py39hd77b12b_0
pyqt5-sip=12.11.0=py39hd77b12b_0
pysocks=1.7.1=py39haa95532_0
python=3.9.17=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
requests=2.29.0=py39haa95532_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py39haa95532_0
sip=6.6.2=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
tornado=6.3.3=py39h2bbff1b_0
tqdm=4.65.0=pypi_0
typing_extensions=4.6.3=py39haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xz=5.4.2=h8cc25b3_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=he774522_0
yapf=0.40.1=pypi_0
zipp=3.16.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
